requirements.txt
snowflake-connector-python>=3.15.0,<4.0.0
snowflake-sqlalchemy>=1.4.0,<2.0.0
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0

